Adopt use of dataset_id as ingestion variable
Support asap-crn-cloud-dataset-metadata repo adoption of dataset_id (like team-smith-pmdbs-sc-rnaseq) as variable name whenever a dataset identifier is passed to functions. Restrict use of dataset_name (like smith-pmdbs-sc-rnaseq) only for outputs or strictly necessary

* Two new modules for housekeeping:  release_util_deprecating.py and summary_stats.py, both of which cleanup release_util.py to keep only activ code there, and make a separate module for summary stats calculation, respectively

* Other housekeeping changes are documented in https://app.clickup.com/t/9014209604/BIOS-2079
